Add MessagePopUp-specific evaluation data in raw assessment page
